Crate objc2_core_text

Bindings to the CoreText framework

See Apple’s docs and the general docs on framework crates for more information.

Structs

CTFont

Apple’s documentation

CTFontDescriptor

Apple’s documentation

CTFontDescriptorMatchingState

Progress state

CTFontFormat

The font is not a recognized format

CTFontOptions

Options for descriptor match and font creation.

CTFontOrientation

Specifies the intended rendering orientation of the font for obtaining glyph metrics.

CTFontTableOptions

Apple’s documentation

CTFontUIFontType

These constants represent the specific user interface purpose to specify for font creation.

Constants

kCTFontPriorityComputer

Apple’s documentation

kCTFontPriorityDynamic

Apple’s documentation

kCTFontPriorityNetwork

Apple’s documentation

kCTFontPriorityProcess

Apple’s documentation

kCTFontPrioritySystem

Apple’s documentation

kCTFontPriorityUser

Apple’s documentation

kCTFontTableAcnt

Apple’s documentation

kCTFontTableAnkr

Apple’s documentation

kCTFontTableAvar

Apple’s documentation

kCTFontTableBASE

Apple’s documentation

kCTFontTableBdat

Apple’s documentation

kCTFontTableBhed

Apple’s documentation

kCTFontTableBloc

Apple’s documentation

kCTFontTableBsln

Apple’s documentation

kCTFontTableCBDT

Apple’s documentation

kCTFontTableCBLC

Apple’s documentation

kCTFontTableCFF

Apple’s documentation

kCTFontTableCFF2

Apple’s documentation

kCTFontTableCOLR

Apple’s documentation

kCTFontTableCPAL

Apple’s documentation

kCTFontTableCidg

Apple’s documentation

kCTFontTableCmap

Apple’s documentation

kCTFontTableCvar

Apple’s documentation

kCTFontTableCvt

Apple’s documentation

kCTFontTableDSIG

Apple’s documentation

kCTFontTableEBDT

Apple’s documentation

kCTFontTableEBLC

Apple’s documentation

kCTFontTableEBSC

Apple’s documentation

kCTFontTableFdsc

Apple’s documentation

kCTFontTableFeat

Apple’s documentation

kCTFontTableFmtx

Apple’s documentation

kCTFontTableFond

Apple’s documentation

kCTFontTableFpgm

Apple’s documentation

kCTFontTableFvar

Apple’s documentation

kCTFontTableGDEF

Apple’s documentation

kCTFontTableGPOS

Apple’s documentation

kCTFontTableGSUB

Apple’s documentation

kCTFontTableGasp

Apple’s documentation

kCTFontTableGlyf

Apple’s documentation

kCTFontTableGvar

Apple’s documentation

kCTFontTableHVAR

Apple’s documentation

kCTFontTableHdmx

Apple’s documentation

kCTFontTableHead

Apple’s documentation

kCTFontTableHhea

Apple’s documentation

kCTFontTableHmtx

Apple’s documentation

kCTFontTableHsty

Apple’s documentation

kCTFontTableJSTF

Apple’s documentation

kCTFontTableJust

Apple’s documentation

kCTFontTableKern

Apple’s documentation

kCTFontTableKerx

Apple’s documentation

kCTFontTableLTSH

Apple’s documentation

kCTFontTableLcar

Apple’s documentation

kCTFontTableLoca

Apple’s documentation

kCTFontTableLtag

Apple’s documentation

kCTFontTableMATH

Apple’s documentation

kCTFontTableMERG

Apple’s documentation

kCTFontTableMVAR

Apple’s documentation

kCTFontTableMaxp

Apple’s documentation

kCTFontTableMeta

Apple’s documentation

kCTFontTableMort

Apple’s documentation

kCTFontTableMorx

Apple’s documentation

kCTFontTableName

Apple’s documentation

kCTFontTableOS2

Apple’s documentation

kCTFontTableOpbd

Apple’s documentation

kCTFontTablePCLT

Apple’s documentation

kCTFontTablePost

Apple’s documentation

kCTFontTablePrep

Apple’s documentation

kCTFontTableProp

Apple’s documentation

kCTFontTableSTAT

Apple’s documentation

kCTFontTableSVG

Apple’s documentation

kCTFontTableSbit

Apple’s documentation

kCTFontTableSbix

Apple’s documentation

kCTFontTableTrak

Apple’s documentation

kCTFontTableVDMX

Apple’s documentation

kCTFontTableVORG

Apple’s documentation

kCTFontTableVVAR

Apple’s documentation

kCTFontTableVhea

Apple’s documentation

kCTFontTableVmtx

Apple’s documentation

kCTFontTableXref

Apple’s documentation

kCTFontTableZapf

Apple’s documentation

kCTVersionNumber10_5

Apple’s documentation

kCTVersionNumber10_6

Apple’s documentation

kCTVersionNumber10_7

Apple’s documentation

kCTVersionNumber10_8

Apple’s documentation

kCTVersionNumber10_9

Apple’s documentation

kCTVersionNumber10_5_2

Apple’s documentation

kCTVersionNumber10_5_3

Apple’s documentation

kCTVersionNumber10_5_5

Apple’s documentation

kCTVersionNumber10_10

Apple’s documentation

kCTVersionNumber10_11

Apple’s documentation

kCTVersionNumber10_12

Apple’s documentation

kCTVersionNumber10_13

Apple’s documentation

kCTVersionNumber10_14

Apple’s documentation

kCTVersionNumber10_15

Apple’s documentation

kCTVersionNumber11_0

Apple’s documentation

Statics

kCTBaselineClassHanging⚠

kCTBaselineClassHanging

kCTBaselineClassIdeographicCentered⚠

kCTBaselineClassIdeographicCentered

kCTBaselineClassIdeographicHigh⚠

kCTBaselineClassIdeographicHigh

kCTBaselineClassIdeographicLow⚠

kCTBaselineClassIdeographicLow

kCTBaselineClassMath⚠

kCTBaselineClassMath

kCTBaselineClassRoman⚠

kCTBaselineClassRoman

kCTBaselineOriginalFont⚠

kCTBaselineOriginalFont

kCTBaselineReferenceFont⚠

kCTBaselineReferenceFont

kCTFontBaselineAdjustAttribute⚠

kCTFontBaselineAdjustAttribute

kCTFontCascadeListAttribute⚠

kCTFontCascadeListAttribute

kCTFontCharacterSetAttribute⚠

kCTFontCharacterSetAttribute

kCTFontCopyrightNameKey⚠

kCTFontCopyrightNameKey

kCTFontDescriptionNameKey⚠

kCTFontDescriptionNameKey

kCTFontDescriptorMatchingCurrentAssetSize⚠

Apple’s documentation

kCTFontDescriptorMatchingDescriptors⚠

Apple’s documentation

kCTFontDescriptorMatchingError⚠

Apple’s documentation

kCTFontDescriptorMatchingPercentage⚠

Apple’s documentation

kCTFontDescriptorMatchingResult⚠

Apple’s documentation

kCTFontDescriptorMatchingSourceDescriptor⚠

Apple’s documentation

kCTFontDescriptorMatchingTotalAssetSize⚠

Apple’s documentation

kCTFontDescriptorMatchingTotalDownloadedSize⚠

Apple’s documentation

kCTFontDesignerNameKey⚠

kCTFontDesignerNameKey

kCTFontDesignerURLNameKey⚠

kCTFontDesignerURLNameKey

kCTFontDisplayNameAttribute⚠

kCTFontDisplayNameAttribute

kCTFontDownloadableAttribute⚠

kCTFontDownloadableAttribute

kCTFontDownloadedAttribute⚠

kCTFontDownloadedAttribute

kCTFontEnabledAttribute⚠

kCTFontEnabledAttribute

kCTFontFamilyNameAttribute⚠

kCTFontFamilyNameAttribute

kCTFontFamilyNameKey⚠

kCTFontFamilyNameKey

kCTFontFeatureSampleTextKey⚠

kCTFontFeatureSampleTextKey

kCTFontFeatureSelectorDefaultKey⚠

kCTFontFeatureSelectorDefaultKey

kCTFontFeatureSelectorIdentifierKey⚠

kCTFontFeatureSelectorIdentifierKey

kCTFontFeatureSelectorNameKey⚠

kCTFontFeatureSelectorNameKey

kCTFontFeatureSelectorSettingKey⚠

kCTFontFeatureSelectorSettingKey

kCTFontFeatureSettingsAttribute⚠

kCTFontFeatureSettingsAttribute

kCTFontFeatureTooltipTextKey⚠

kCTFontFeatureTooltipTextKey

kCTFontFeatureTypeExclusiveKey⚠

kCTFontFeatureTypeExclusiveKey

kCTFontFeatureTypeIdentifierKey⚠

kCTFontFeatureTypeIdentifierKey

kCTFontFeatureTypeNameKey⚠

kCTFontFeatureTypeNameKey

kCTFontFeatureTypeSelectorsKey⚠

kCTFontFeatureTypeSelectorsKey

kCTFontFeaturesAttribute⚠

kCTFontFeaturesAttribute

kCTFontFixedAdvanceAttribute⚠

kCTFontFixedAdvanceAttribute

kCTFontFormatAttribute⚠

kCTFontFormatAttribute

kCTFontFullNameKey⚠

kCTFontFullNameKey

kCTFontLanguagesAttribute⚠

kCTFontLanguagesAttribute

kCTFontLicenseNameKey⚠

kCTFontLicenseNameKey

kCTFontLicenseURLNameKey⚠

kCTFontLicenseURLNameKey

kCTFontMacintoshEncodingsAttribute⚠

kCTFontMacintoshEncodingsAttribute

kCTFontManufacturerNameKey⚠

kCTFontManufacturerNameKey

kCTFontMatrixAttribute⚠

kCTFontMatrixAttribute

kCTFontNameAttribute⚠

kCTFontNameAttribute

kCTFontOpenTypeFeatureTag⚠

kCTFontOpenTypeFeatureTag

kCTFontOpenTypeFeatureValue⚠

kCTFontOpenTypeFeatureValue

kCTFontOpticalSizeAttribute⚠

kCTFontOpticalSizeAttribute

kCTFontOrientationAttribute⚠

kCTFontOrientationAttribute

kCTFontPostScriptCIDNameKey⚠

kCTFontPostScriptCIDNameKey

kCTFontPostScriptNameKey⚠

kCTFontPostScriptNameKey

kCTFontPriorityAttribute⚠

kCTFontPriorityAttribute

kCTFontRegistrationScopeAttribute⚠

kCTFontRegistrationScopeAttribute

kCTFontSampleTextNameKey⚠

kCTFontSampleTextNameKey

kCTFontSizeAttribute⚠

kCTFontSizeAttribute

kCTFontStyleNameAttribute⚠

kCTFontStyleNameAttribute

kCTFontStyleNameKey⚠

kCTFontStyleNameKey

kCTFontSubFamilyNameKey⚠

kCTFontSubFamilyNameKey

kCTFontTrademarkNameKey⚠

kCTFontTrademarkNameKey

kCTFontTraitsAttribute⚠

kCTFontTraitsAttribute

kCTFontURLAttribute⚠

kCTFontURLAttribute

kCTFontUniqueNameKey⚠

kCTFontUniqueNameKey

kCTFontVariationAttribute⚠

kCTFontVariationAttribute

kCTFontVariationAxesAttribute⚠

kCTFontVariationAxesAttribute

kCTFontVariationAxisDefaultValueKey⚠

kCTFontVariationAxisDefaultValueKey

kCTFontVariationAxisHiddenKey⚠

kCTFontVariationAxisHiddenKey

kCTFontVariationAxisIdentifierKey⚠

kCTFontVariationAxisIdentifierKey

kCTFontVariationAxisMaximumValueKey⚠

kCTFontVariationAxisMaximumValueKey

kCTFontVariationAxisMinimumValueKey⚠

kCTFontVariationAxisMinimumValueKey

kCTFontVariationAxisNameKey⚠

kCTFontVariationAxisNameKey

kCTFontVendorURLNameKey⚠

kCTFontVendorURLNameKey

kCTFontVersionNameKey⚠

kCTFontVersionNameKey

Functions

CTFontCopyAttributeDeprecated

CTFontCopyAvailableTablesDeprecated

CTFontCopyCharacterSetDeprecated

CTFontCopyDefaultCascadeListForLanguages⚠Deprecated

CTFontCopyDisplayNameDeprecated

CTFontCopyFamilyNameDeprecated

CTFontCopyFeatureSettingsDeprecated

CTFontCopyFeaturesDeprecated

CTFontCopyFontDescriptorDeprecated

CTFontCopyFullNameDeprecated

CTFontCopyLocalizedName⚠Deprecated

CTFontCopyNameDeprecated

CTFontCopyPostScriptNameDeprecated

CTFontCopySupportedLanguagesDeprecated

CTFontCopyTableDeprecated

CTFontCopyTraitsDeprecated

CTFontCopyVariationDeprecated

CTFontCopyVariationAxesDeprecated

CTFontCreateCopyWithAttributes⚠Deprecated

CTFontCreateCopyWithFamily⚠Deprecated

CTFontCreateForStringDeprecated

CTFontCreateForStringWithLanguageDeprecated

CTFontCreateUIFontForLanguageDeprecated

CTFontCreateWithFontDescriptor⚠Deprecated

CTFontCreateWithFontDescriptorAndOptions⚠Deprecated

CTFontCreateWithName⚠Deprecated

CTFontCreateWithNameAndOptions⚠Deprecated

CTFontCreateWithQuickdrawInstance⚠Deprecated

CTFontDescriptorCopyAttributeDeprecated

CTFontDescriptorCopyAttributesDeprecated

CTFontDescriptorCopyLocalizedAttribute⚠Deprecated

CTFontDescriptorCreateCopyWithAttributes⚠Deprecated

CTFontDescriptorCreateCopyWithFamilyDeprecated

CTFontDescriptorCreateCopyWithFeatureDeprecated

CTFontDescriptorCreateCopyWithVariationDeprecated

CTFontDescriptorCreateMatchingFontDescriptor⚠Deprecated

CTFontDescriptorCreateMatchingFontDescriptors⚠Deprecated

CTFontDescriptorCreateWithAttributes⚠Deprecated

CTFontDescriptorCreateWithNameAndSizeDeprecated

CTFontGetAscentDeprecated

CTFontGetBoundingBoxDeprecated

CTFontGetCapHeightDeprecated

CTFontGetDescentDeprecated

CTFontGetGlyphCountDeprecated

CTFontGetLeadingDeprecated

CTFontGetMatrixDeprecated

CTFontGetSizeDeprecated

CTFontGetSlantAngleDeprecated

CTFontGetStringEncodingDeprecated

CTFontGetUnderlinePositionDeprecated

CTFontGetUnderlineThicknessDeprecated

CTFontGetUnitsPerEmDeprecated

CTFontGetXHeightDeprecated

CTFontHasTableDeprecated

CTGetCoreTextVersionDeprecated

Returns the version of the CoreText framework.

Type Aliases

ATSFontRef

Apple’s documentation

CTFontPriority

Apple’s documentation

CTFontTableTag

Apple’s documentation