objc2_core_text

Struct CTFont

pub struct CTFont { /* private fields */ }

Apple’s documentation

This is toll-free bridged with NSFont.

Implementations

impl CTFont

pub unsafe fn with_name( name: &CFString, size: CGFloat, matrix: *const CGAffineTransform, ) -> CFRetained<CTFont>

Returns a new font reference for the given name.

This function uses font descriptor matching so only registered fonts can be returned; see CTFontManager.h for more information. If you are trying to create a system UI font (with name beginning with a “.”), you should use CTFontCreateUIFontForLanguage() or appropriate AppKit/UIKit APIs instead.

Parameter name: The font name for which you wish to create a new font reference. A valid PostScript name is preferred, although other font name types will be matched in a fallback manner. Any font name beginning with a “.” is reserved for the system and should not be used here.

Parameter size: The point size for the font reference. If 0.0 is specified, the default font size of 12.0 will be used.

Parameter matrix: The transformation matrix for the font. If unspecified, the identity matrix will be used. Optional.

Returns: This function will return a CTFontRef that best matches the name provided with size and matrix attributes. The name parameter is the only required parameters, and default values will be used for unspecified parameters. A best match will be found if all parameters cannot be matched identically.

Safety

matrix must be a valid pointer or null.

pub unsafe fn with_font_descriptor( descriptor: &CTFontDescriptor, size: CGFloat, matrix: *const CGAffineTransform, ) -> CFRetained<CTFont>

Returns a new font reference that best matches the font descriptor.

Parameter descriptor: A font descriptor containing attributes that specify the requested font.

Parameter size: The point size for the font reference. If 0.0 is specified and the font descriptor does not specify the size, the default font size of 12.0 will be used.

Parameter matrix: The transformation matrix for the font. If unspecified, the identity matrix will be used. Optional.

Returns: This function will return a CTFontRef that best matches the attributes provided with the font descriptor. The size and matrix parameters will override any specified in the font descriptor, unless they are unspecified. A best match font will always be returned, and default values will be used for any unspecified.

Safety

matrix must be a valid pointer or null.

impl CTFont

pub unsafe fn with_name_and_options( name: &CFString, size: CGFloat, matrix: *const CGAffineTransform, options: CTFontOptions, ) -> CFRetained<CTFont>

Returns a new font reference for the given name.

This function uses font descriptor matching so only registered fonts can be returned; see CTFontManager.h for more information. If you are trying to create a system UI font (with name beginning with a “.”), you should use CTFontCreateUIFontForLanguage() or appropriate AppKit/UIKit APIs instead.

Parameter name: The font name for which you wish to create a new font reference. A valid PostScript name is preferred, although other font name types will be matched in a fallback manner. Any font name beginning with a “.” is reserved for the system and should not be used here.

Parameter size: The point size for the font reference. If 0.0 is specified, the default font size of 12.0 will be used.

Parameter matrix: The transformation matrix for the font. If unspecified, the identity matrix will be used. Optional.

Parameter options: Options flags.

Returns: This function will return a CTFontRef that best matches the name provided with size and matrix attributes. The name parameter is the only required parameters, and default values will be used for unspecified parameters. A best match will be found if all parameters cannot be matched identically.

Safety

matrix must be a valid pointer or null.

pub unsafe fn with_font_descriptor_and_options( descriptor: &CTFontDescriptor, size: CGFloat, matrix: *const CGAffineTransform, options: CTFontOptions, ) -> CFRetained<CTFont>

Returns a new font reference that best matches the font descriptor.

Parameter descriptor: A font descriptor containing attributes that specify the requested font.

Parameter size: The point size for the font reference. If 0.0 is specified, the default font size of 12.0 will be used.

Parameter matrix: The transformation matrix for the font. If unspecified, the identity matrix will be used. Optional.

Parameter options: Options flags.

Returns: This function will return a CTFontRef that best matches the attributes provided with the font descriptor. The size and matrix parameters will override any specified in the font descriptor, unless they are unspecified. A best match font will always be returned, and default values will be used for any unspecified.

Safety

matrix must be a valid pointer or null.

impl CTFont

pub fn new_ui_font_for_language( ui_type: CTFontUIFontType, size: CGFloat, language: Option<&CFString>, ) -> Option<CFRetained<CTFont>>

Returns the special UI font for the given language and UI type.

Parameter uiType: A uiType constant specifying the intended UI use for the requested font reference.

Parameter size: The point size for the font reference. If 0.0 is specified, the default size for the requested uiType is used.

Parameter language: Language identifier to select a font for a particular localization. If unspecified, the current system language is used. The format of the language identifier should conform to UTS #35.

Returns: This function returns the correct font for various UI uses. The only required parameter is the uiType selector, unspecified optional parameters will use default values.

pub unsafe fn copy_with_attributes( &self, size: CGFloat, matrix: *const CGAffineTransform, attributes: Option<&CTFontDescriptor>, ) -> CFRetained<CTFont>

Returns a new font with additional attributes based on the original font.

This function provides a mechanism to quickly change attributes on a given font reference in response to user actions. For instance, the size can be changed in response to a user manipulating a size slider.

Parameter font: Original font reference to base new font on.

Parameter size: The point size for the font reference. If 0.0 is specified, the original font’s size will be preserved.

Parameter matrix: The transformation matrix for the font. If unspecified, the original font matrix will be preserved. Optional.

Parameter attributes: A font descriptor containing additional attributes that the new font should contain.

Returns: Returns a new font reference converted from the original with the specified attributes.

Safety

matrix must be a valid pointer or null.

pub unsafe fn copy_with_family( &self, size: CGFloat, matrix: *const CGAffineTransform, family: &CFString, ) -> Option<CFRetained<CTFont>>

Returns a new font in the specified family based on the traits of the original font.

Parameter font: Original font reference to base new font on.

Parameter size: The point size for the font reference. If 0.0 is specified, the original font’s size will be preserved.

Parameter matrix: The transformation matrix for the font. If unspecified, the original font matrix will be preserved. Optional.

Parameter family: The name of the desired family.

Returns: Returns a new font reference with the original traits in the given family. NULL if non found in the system.

Safety

matrix must be a valid pointer or null.

pub fn for_string( &self, string: &CFString, range: CFRange, ) -> CFRetained<CTFont>

Returns a new font reference that can best map the given string range based on the current font.

This function is to be used when the current font does not cover the given range of the string. The current font itself will not be returned, but preference is given to fonts in its cascade list.

Parameter currentFont: The current font that contains a valid cascade list.

Parameter string: A unicode string containing characters that cannot be encoded by the current font.

Parameter range: A CFRange specifying the range of the string that needs to be mapped.

Returns: This function returns the best substitute font that can encode the specified string range.

See also: CTFontCopyCharacterSet

See also: CTFontGetGlyphsForCharacters

See also: kCTFontCascadeListAttribute

pub fn for_string_with_language( &self, string: &CFString, range: CFRange, language: Option<&CFString>, ) -> CFRetained<CTFont>

Returns a new font reference that can best map the given string range based on the current font and language specified.

The current font itself can be returned if it covers the string provided.

Parameter currentFont: The current font that contains a valid cascade list.

Parameter string: A unicode string containing characters that cannot be encoded by the current font.

Parameter range: A CFRange specifying the range of the string that needs to be mapped.

Parameter language: Language identifier to select a font for a particular localization. If unspecified, the current system language is used. The format of the language identifier should conform to UTS #35.

Returns: This function returns the best substitute font that can encode the specified string range.

See also: CTFontCopyCharacterSet

See also: CTFontGetGlyphsForCharacters

See also: kCTFontCascadeListAttribute

pub fn font_descriptor(&self) -> CFRetained<CTFontDescriptor>

Returns the normalized font descriptors for the given font reference.

Parameter font: The font reference.

Returns: This function returns a normalized font descriptor for a font. The font descriptor contains enough information to recreate this font at a later time.

pub fn attribute(&self, attribute: &CFString) -> Option<CFRetained<CFType>>

Returns the value associated with an arbitrary attribute.

Parameter font: The font reference.

Parameter attribute: The requested attribute.

Returns: This function returns a retained reference to an arbitrary attribute. If the requested attribute is not present, NULL is returned. Refer to the attribute definitions for documentation as to how each attribute is packaged as a CFType.

pub fn size(&self) -> CGFloat

Returns the point size of the font reference.

Parameter font: The font reference.

Returns: This function returns the point size of the given font reference. This is the point size provided when the font was created.

pub fn matrix(&self) -> CGAffineTransform

Returns the transformation matrix of the font.

Parameter font: The font reference.

Returns: This function returns the transformation matrix for this given font reference. This is the matrix that was provided when the font was created.

pub fn traits(&self) -> CFRetained<CFDictionary>

Returns the font traits dictionary.

Parameter font: The font reference.

Returns: This function returns a retained reference to the font traits dictionary. Individual traits can be accessed with the trait key constants. See CTFontTraits.h for a definition of the font traits.

pub unsafe fn default_cascade_list_for_languages( &self, language_pref_list: Option<&CFArray>, ) -> Option<CFRetained<CFArray>>

Return an ordered list of CTFontDescriptorRef’s for font fallback derived from the system default fallback according to the given language preferences, making a reasonable attempt to match the given font’s style, weight, and width.

Parameter font: The font reference.

Parameter languagePrefList: An array of language identifiers as CFString values, in decreasing order of preference.

Returns: The ordered list of fallback fonts - ordered array of CTFontDescriptors.

Safety

language_pref_list generic must be of the correct type.

pub fn post_script_name(&self) -> CFRetained<CFString>

Returns the PostScript name.

Parameter font: The font reference.

Returns: This function returns a retained reference to the PostScript name of the font.

pub fn family_name(&self) -> CFRetained<CFString>

Returns the family name.

Parameter font: The font reference.

Returns: This function returns a retained reference to the family name of the font.

pub fn full_name(&self) -> CFRetained<CFString>

Returns the display name.

Parameter font: The font reference.

Returns: This function returns a retained reference to the full name of the font.

pub fn display_name(&self) -> CFRetained<CFString>

Returns the display name.

Parameter font: The font reference.

Returns: This function returns a retained reference to the localized display name of the font.

pub fn name(&self, name_key: &CFString) -> Option<CFRetained<CFString>>

Returns a reference to the requested name.

Parameter font: The font reference.

Parameter nameKey: The name specifier. See name specifier constants.

Returns: This function creates the requested name for the font, or NULL if the font does not have an entry for the requested name. The Unicode version of the name will be preferred, otherwise the first available will be used.

pub unsafe fn localized_name( &self, name_key: &CFString, actual_language: *mut *const CFString, ) -> Option<CFRetained<CFString>>

Returns a reference to a localized font name.

Localized names are necessary for presentation to a human but are rarely appropriate for programmatic use. CoreText provides localizations for common names but will not attempt any sort of automated translation.

Parameter font: The font reference.

Parameter nameKey: A name specifier listed in “Font Constants”, for example kCTFontStyleNameKey. Name keys present in dictionaries (such as those for axes or features) are handled not by this function but by the functions returning those dictionaries.

Parameter actualLanguage: Pointer to a CFStringRef to receive the language identifier of the returned name string. The format of the language identifier will conform to UTS #35. If CoreText can supply its own localized string where the font cannot, this value will be NULL.

Returns: This function returns a specific localized name from the font reference. The name is localized based on the user’s global language precedence. If the font does not have an entry for the requested name, NULL will be returned. The matched language will be returned in the caller’s buffer.

See also: CTFontCopyVariationAxes

See also: CTFontCopyFeatures

Safety

actual_language must be a valid pointer or null.

pub fn character_set(&self) -> CFRetained<CFCharacterSet>

Returns the Unicode character set of the font.

Parameter font: The font reference.

Returns: This function returns a retained reference to the font’s character set. This character set covers the nominal referenced by the font’s Unicode cmap table (or equivalent).

pub fn string_encoding(&self) -> CFStringEncoding

Returns the best string encoding for legacy format support.

Parameter font: The font reference.

Returns: This function returns the best string encoding for the font.

pub fn supported_languages(&self) -> CFRetained<CFArray>

Returns an array of languages supported by the font.

Parameter font: The font reference.

Returns: This function returns a retained reference to an array of languages supported by the font. The array contains language identifier strings as CFStringRefs. The format of the language identifier will conform to UTS #35.

pub fn ascent(&self) -> CGFloat

Returns the scaled font ascent metric.

Parameter font: The font reference.

Returns: This function returns the font ascent metric scaled based on the point size and matrix of the font reference.

pub fn descent(&self) -> CGFloat

Returns the scaled font descent metric.

Parameter font: The font reference.

Returns: This function returns the font descent metric scaled based on the point size and matrix of the font reference.

pub fn leading(&self) -> CGFloat

Returns the scaled font leading metric.

Parameter font: The font reference.

Returns: This function returns the font leading metric scaled based on the point size and matrix of the font reference.

pub fn units_per_em(&self) -> c_uint

Returns the units per em metric.

Parameter font: The font reference.

Returns: This function returns the units per em of the font.

pub fn glyph_count(&self) -> CFIndex

Returns the number of glyphs.

Parameter font: The font reference.

Returns: This function returns the number of glyphs in the font.

pub fn bounding_box(&self) -> CGRect

Returns the scaled bounding box.

Parameter font: The font reference.

Returns: This will return the design bounding box of the font, which is the rectangle defined by xMin, yMin, xMax, and yMax values for the font.

pub fn underline_position(&self) -> CGFloat

Returns the scaled underline position.

Parameter font: The font reference.

Returns: This function returns the font underline position metric scaled based on the point size and matrix of the font reference.

pub fn underline_thickness(&self) -> CGFloat

Returns the scaled underline thickness metric.

Parameter font: The font reference.

Returns: This function returns the font underline thickness metric scaled based on the point size and matrix of the font reference.

pub fn slant_angle(&self) -> CGFloat

Returns the slant angle of the font.

Parameter font: The font reference.

Returns: This function returns the transformed slant angle of the font. This is equivalent to the italic or caret angle with any skew from the transformation matrix applied.

pub fn cap_height(&self) -> CGFloat

Returns the cap height metric.

Parameter font: The font reference.

Returns: This function returns the font cap height metric scaled based on the point size and matrix of the font reference.

pub fn x_height(&self) -> CGFloat

Returns the X height metric.

Parameter font: The font reference.

Returns: This function returns the font X height metric scaled based on the point size and matrix of the font reference.

impl CTFont

pub fn variation_axes(&self) -> Option<CFRetained<CFArray>>

Returns an array of variation axis dictionaries.

Each variation axis dictionary contains the five kCTFontVariationAxis* keys above, and kCTFontVariationAxisNameKey values will be localized when supported by the font; for programmatic uses kCTFontVariationAxesAttribute may be used instead.

Parameter font: The font reference.

Returns: An array of variation axis dictionaries or null if the font does not support variations.

See also: kCTFontVariationAxesAttribute

pub fn variation(&self) -> Option<CFRetained<CFDictionary>>

Returns a variation dictionary.

This function describes the current configuration of a variation font: a dictionary of number values with variation identifier number keys. As of macOS 10.12 and iOS 10.0, only non-default values (as determined by the variation axis) are returned.

Parameter font: The font reference.

Returns: This function returns a variation dictionary or null if the font does not support variations.

See also: kCTFontVariationAxisIdentifierKey

See also: kCTFontVariationAxisDefaultValueKey

impl CTFont

pub fn features(&self) -> Option<CFRetained<CFArray>>

Returns an array of font features

The returned value describes the features available for the provided font. Each array value is a feature dictionary describing a feature type, with related selector dictionaries in an array under the kCTFontFeatureTypeSelectorsKey. While CoreText supports AAT and OpenType font features, they are preferentially represented as AAT features owing to their more formal structure: individual feature types can be either exclusive or non-exclusive, which indicates whether one or more of its selectors can be simultaneously enabled. Where possible features are elaborated with their OpenType feature tag and value, which can occur within both type or selector dictionaries depending on the feature’s mapping to an AAT type and selector pair. Names are localized according to the preferred langauges of the caller and therefore are not appropriate for programmatically identifying features.

Parameter font: The font reference.

Returns: This function returns an array of font feature dictionaries for the font reference.

pub fn feature_settings(&self) -> Option<CFRetained<CFArray>>

Returns an array of font feature setting tuples

A setting tuple is a dictionary of a kCTFontFeatureTypeIdentifierKey key-value pair and a kCTFontFeatureSelectorIdentifierKey key-value pair. Each tuple corresponds to an enabled non-default setting. It is the caller’s responsibility to handle exclusive and non-exclusive settings as necessary.

Parameter font: The font reference.

Returns: This function returns a normalized array of font feature setting dictionaries. The array will only contain the non-default settings that should be applied to the font, or NULL if the default settings should be used.

impl CTFont

pub unsafe fn with_quickdraw_instance( name: *const c_char, identifier: i16, style: u8, size: CGFloat, ) -> CFRetained<CTFont>

👎Deprecated: Quickdraw font references are deprecated

Returns a font reference for the given Quickdraw instance.

This function is provided for compatibility support between Core Text and clients needing to support Quickdraw font references.

Parameter name: The Quickdraw font name. If NULL or zero length, an identifier must be specified instead.

Parameter identifier: The Quickdraw font identifier. If 0, a name must be specified instead.

Parameter style: The Quickdraw font style.

Parameter size: The point size for the font reference. If 0.0 is specified, the default size of 12.0 is used.

Returns: This function returns the best font instance matching the Quickdraw instance information.

Safety

name must be a valid pointer or null.

impl CTFont

pub fn available_tables( &self, options: CTFontTableOptions, ) -> Option<CFRetained<CFArray>>

Returns an array of font table tags.

Parameter font: The font reference.

Parameter options: The options used when copying font tables.

Returns: This function returns an array of CTFontTableTag values for the given font and the supplied options. The returned set will contain unboxed values, which may be extracted like so: CTFontTableTag tag = (CTFontTableTag)(uintptr_t)CFArrayGetValueAtIndex(tags, index);

pub fn has_table(&self, tag: CTFontTableTag) -> bool

Determine whether a table is present in a font.

This is a convenience function to avoid requesting an array of table tags or any table data. It behaves according as if using kCTFontTableOptionNoOptions.

Parameter font: The font reference.

Parameter tag: The font table identifier as a CTFontTableTag.

Returns: Returns true if the call was successful and the requested table is present.

pub fn table( &self, table: CTFontTableTag, options: CTFontTableOptions, ) -> Option<CFRetained<CFData>>

Returns a reference to the font table data.

Parameter font: The font reference.

Parameter table: The font table identifier as a CTFontTableTag.

Parameter options: The options used when copying font table.

Returns: This function returns a retained reference to the font table data as CFDataRef or NULL if the table is not present.